@@ -26,7 +26,7 @@ class UsersController < ApplicationController
     end
 
     def user_params
-        params.permit(:username, :password)
+        params.require(:user).permit(:username, :password)
     end
 
 
