@@ -11,8 +11,8 @@ function Navbar () {
   <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
     <div class="navbar-nav">
       <a class="nav-item nav-link active" href="/">Home <span class="sr-only">(current)</span></a>
-      <a class="nav-item nav-link" href="/Forsale">Mountains</a>
-      <a class="nav-item nav-link" href="/showCars">Cars for sale</a>
+      <a class="nav-item nav-link" href="/mountains">Mountains</a>
+      {/* <a class="nav-item nav-link" href="/showCars">Cars for sale</a> */}
     </div>
   </div>
 </nav>
